Methods the create, find and delete finished.

+++ src/main/kotlin/com/br/project/repositories/VeiculoRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface VeiculoRepository : CrudRepository<VeiculoModel, Long>
+interface VeiculoRepository : CrudRepository<VeiculoModel, Int>
